+++ devices/temperature_humidity_sensor.py
@@ -1,0 +1,13 @@
+import Domoticz
+from devices.device import Device
+
+
+class TemperatureHumiditySensor(Device):
+    def create_device(self, unit, device_id, device_name):
+        return Domoticz.Device(Unit=unit, DeviceID=device_id, Name=device_name, TypeName="Temp+Hum").Create()
+
+    def get_numeric_value(self, value, device):
+        return 0
+
+    def get_string_value(self, value, device):
+        return str(value)